llm.py
from dotenv import load_dotenv
from pydantic import BaseModel
from langchain_openai import ChatOpenAI
from langchain_anthropic import ChatAnthropic
from langchain_core.prompts import ChatPromptTemplate
from langchain_core.output_parsers import PydanticOutputParser
from langchain_core.agents import AgentExecutor
from tools import search_tool, wiki_tool

load_dotenv()

tools = [search_tool, wiki_tool]
class ResearchResponse(BaseModel):
    topic: str
    summary: str
    sources: list[str]
    tools_used: list[str]

llm = ChatAnthropic(model="claude-3-5-sonnet")
parser = PydanticOutputParser(pydantic_object=ResearchResponse)

prompt = ChatPromptTemplate.from_messages(
    [
        (
            "system", 
            """
            You are a helpful research assistant that will help generate a research paper.
            Answer the user query and use necessary tools.
            Wrap the output in this format and provide no other text\n{format_instructions}
            """,
        ),
        ('placeholder', '{chat_history}'),
        ('human', '{query}'),
        ('placeholder', '{agent_scratchpad}'),

    ]
).partial(format_instructions=parser.get_format_instructions())

response = llm.invoke('What is the meaning of life?')
agent = create_tool_calling_agent(
    llm=llm,
    prompt=prompt,
    tools=tools
)

agent_executor = AgentExecutor(agent=agent, tools=[], verbose=True)
query = input('what can i help you research?')
raw_response = agent_executor.invoke({'query': query, 'name': 'Alex'})

print(raw_response)

try:
    structured_response = parser.parse(raw_response.get('output')[0]['text'])
except Exception as e:
    print(f"Error parsing response: {e}")
    structured_response = None
print(structured_response.topic)